Restructuring repository
* All current code moved to pg_tde subdirectory
* Makefile adjusted
* Meson build file added
* Gitignore added to ignore in-source files with make
