+++ .gitignore
@@ -1,0 +1,2 @@
+pg_tde.so
+*.o
