+++ Makefile
@@ -1,27 +1,17 @@
 # contrib/pg_tde/Makefile
 
 MODULE_big = pg_tde
-#OBJS = \
-        pg_tdeam.o \
-        pg_tdeam_handler.o \
-        pg_tdeam_visibility.o \
-        pg_tdetoast.o \
-        pg_tde_io.o \
-        pg_tde_prune.o \
-        pg_tde_rewrite.o \
-        pg_tde_vacuumlazy.o \
-        pg_tde_visibilitymap.o
 
 OBJS = \
-        pg_tdeam.o \
-        pg_tdeam_handler.o \
-        pg_tdeam_visibility.o \
-        pg_tdetoast.o \
-        pg_tde_io.o \
-        pg_tde_prune.o \
-        pg_tde_rewrite.o \
-        pg_tde_vacuumlazy.o \
-        pg_tde_visibilitymap.o
+        pg_tde/pg_tdeam.o \
+        pg_tde/pg_tdeam_handler.o \
+        pg_tde/pg_tdeam_visibility.o \
+        pg_tde/pg_tdetoast.o \
+        pg_tde/pg_tde_io.o \
+        pg_tde/pg_tde_prune.o \
+        pg_tde/pg_tde_rewrite.o \
+        pg_tde/pg_tde_vacuumlazy.o \
+        pg_tde/pg_tde_visibilitymap.o
 
 EXTENSION = pg_tde
 DATA = pg_tde--1.0.sql
