+++ pg_tde/pg_tde_vacuumlazy.c
@@ -34,9 +34,9 @@
 
 #include <math.h>
 
-#include "pg_tdeam.h"
-#include "pg_tdeam_xlog.h"
-#include "pg_tde_visibilitymap.h"
+#include "pg_tde/pg_tdeam.h"
+#include "pg_tde/pg_tdeam_xlog.h"
+#include "pg_tde/pg_tde_visibilitymap.h"
 
 #include "access/amapi.h"
 #include "access/genam.h"
